requirements.txt
aniso8601==9.0.1
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
coverage==7.3.2
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Autodoc==0.1.2
Flask-RESTful==0.3.10
flask-selfdoc==1.4.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.2
pluggy==1.3.0
pytest==7.4.3
pytest-cov==4.1.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tomli==2.0.1
urllib3==2.0.7
vcrpy==5.1.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.2
